+++ src/bin/hnefatafl-client/main.rs
@@ -488,10 +488,16 @@ fn pass_messages() -> impl Stream<Item = Message> {
                     handle_error(socket.set_tcp_keepalive(&keep_alive));
 
                     if let Err(error) = socket.connect(&address) {
-                        error!("socket.connect {address_string}: {error}");
+                        if error.kind() == ErrorKind::NetworkUnreachable {
+                            handle_error(executor::block_on(
+                                sender.send(Message::TcpConnectFailed),
+                            ));
+                        } else {
+                            handle_error(executor::block_on(sender.send(Message::ServerShutdown)));
+                        }
 
+                        error!("socket.connect {address_string}: {error}");
                         handle_error(executor::block_on(sender.send(Message::TcpDisconnect)));
-                        handle_error(executor::block_on(sender.send(Message::TcpConnectFailed)));
 
                         continue 'start_over;
                     }